Design: CancelConfirmModal UI 완료
